Patched Electron version to fix *COMPACT cursor
The patched version is called ELK100P.

Introduced the PATCH_COMPACT_CURSOR option which patches *COMPACT to
hide the cursor when run and re-enable it afterwards.

There wasn't any space for this, as things stood, but it has been freed
up by removing the arguments to *COMPACT so it always uses the screen
memory and nothing else (i.e. you can't specify your own memory range
to use).

This means that it never tramples on language space and gets more
workspace if you switch to MODEs 0, 1 or 2, so is probably a little
more friendly.

There are still some remnants of the arguments in the argument_-
string_table which should be removed but I haven't worked out exactly
how everything works around that to feel confident removing them.