Patch Electron version renumbered 1.00 -> 1.30E
Also, updated the binary version byte at &8008 from &00 to &30, to
bring it in line with the Model B version 1.30.

Updated these because the AP6 Plus 1 support ROM for the Electron
checks this byte to determine if the ZYSysHelp file needs writing,
which I assume it doesn't as this bug should have been fixed here.

Changed the version string to 1.30E to indicate it's not the original
Electron version but a patched version of 1.30 from the Model B.